changing skips for tests

+++ pyccel/stdlib/ndarrays/ndarrays.h
@@ -142,9 +142,9 @@ int64_t     *numpy_to_ndarray_shape(int64_t *np_shape, int nd);
 
 /* numpy sum */
 
-int32_t            numpy_sum_bool(t_ndarray arr);
-int32_t             numpy_sum_int8(t_ndarray arr);
-int32_t            numpy_sum_int16(t_ndarray arr);
+int64_t            numpy_sum_bool(t_ndarray arr);
+int64_t            numpy_sum_int8(t_ndarray arr);
+int64_t            numpy_sum_int16(t_ndarray arr);
 int64_t            numpy_sum_int32(t_ndarray arr);
 int64_t            numpy_sum_int64(t_ndarray arr);
 float              numpy_sum_float32(t_ndarray arr);
